adding last changes to the chunk_list, nasty stuff, but not more time available for this :(

+++ chunklist/chunk.h
@@ -62,7 +62,7 @@ public:
 
 private:
     alignas (T) T m_data[SIZE];
-    unsigned int m_size;
+    short m_size; // could be a short
 };
